requirements.txt
configparser
mysql-connector
pyyaml
requests
boto3
